requirements.txt
alembic==1.7.7
attrs==21.4.0
boto3==1.22.4
botocore==1.25.4
click==8.1.3
Flask==2.1.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
nose==1.3.7
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
waitress==2.1.1
Werkzeug==2.1.2
zipp==3.8.0
